fix: correct metrics calculation, float precision bug, and scenario label errors

+++ src/config/tool_config.py
@@ -6,7 +6,7 @@ TOOL_CONFIG = {
         "required": ["amount"],
         "rules": {
             "amount": {
-                "type": int,
+                "type": float,
                 "min": 0.01,
                 "max": 10000,
             }
